+++ gpt_ui/src/cli/config.rs
@@ -1,13 +1,14 @@
+use super::dialog::{input, password, select};
 use super::error::Result;
-use dialoguer::{theme::ColorfulTheme, Input, Password, Select};
 use gpt_core::config::Config;
 
 pub(crate) fn config() -> Result<()> {
-    let selection = Select::with_theme(&ColorfulTheme::default())
-        .with_prompt("Action to your config?")
-        .items(&["Display", "Modify"])
-        .interact()?;
-    match selection {
+    #[cfg(all(feature = "dialoguer", not(feature = "cliclack")))]
+    let items = &["Display", "Modify"];
+    #[cfg(all(feature = "cliclack", not(feature = "dialoguer")))]
+    let items = &[(0, "Display", ""), (1, "Modify", "")];
+    let chosen = select("Action to your config?", items)?;
+    match chosen {
         0 => display(),
         1 => modify(),
         _ => unreachable!(),
@@ -28,20 +29,11 @@ fn display() -> Result<()> {
 
 fn modify() -> Result<()> {
     let mut config = Config::read_masked()?;
-    let endpoint = Input::<String>::with_theme(&ColorfulTheme::default())
-        .with_prompt(format!(
-            " Endpoint? Current: {} [Empty to unchange]\n",
-            config.endpoint()
-        ))
-        .allow_empty(true)
-        .interact_text()?;
+    let endpoint = input(" Endpoint? [Empty to unchange]\n", config.endpoint())?;
     if !endpoint.is_empty() {
         config.set_endpoint(endpoint);
     }
-    let api_key = Password::with_theme(&ColorfulTheme::default())
-        .with_prompt("API Key? [Hidden]\n")
-        .allow_empty_password(true)
-        .interact()?;
+    let api_key = password("API Key? [Hidden]\n")?;
     if !api_key.is_empty() {
         config.set_api_key(api_key);
     }