+++ gpt_ui/src/lib.rs
@@ -1,3 +1,8 @@
+#[cfg(all(feature = "dialoguer", feature = "cliclack"))]
+compile_error!("Only one of the features `dialoguer` and `cliclack` can be enabled at a time.");
+#[cfg(not(any(feature = "dialoguer", feature = "cliclack")))]
+compile_error!("At least one of the features `dialoguer` and `cliclack` must be enabled.");
+
 #[cfg(feature = "cli")]
 mod cli;
 #[cfg(feature = "tui")]
@@ -18,3 +23,6 @@ const VERSION: &str = const_format::formatcp!(
     env!("VERGEN_RUSTC_CHANNEL"),
     env!("VERGEN_RUSTC_SEMVER")
 );
+
+#[cfg(feature = "mock")]
+const MOCK_SERVER: &str = "http://127.0.0.1:3000";