+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 ## [0.0.1-alpha3] - 2024-05-26
 
 - fix: config and data dir coflict on windows
+- fix: ctrl-c cannot work on windows
+
 
 ## [0.0.1-alpha2] - 2024-05-26